+++ src/assets/poems/poem2
@@ -1,0 +1,1 @@
+hello i am poem 2
